+++ .storybook/pages/CoursePlannerEdit/index.ts
@@ -1,1 +1,0 @@
-export { CoursePlannerEdit as default } from './CoursePlannerEdit';