+++ src/components/ButtonDropdown/index.ts
@@ -1,1 +1,0 @@
-export { ButtonDropdown as default } from './ButtonDropdown';