+++ src/components/Table/Table.stories.tsx
@@ -1,18 +1,23 @@
 import type { StoryObj, Meta } from '@storybook/react';
 import React, { useState } from 'react';
 
-import { Table } from './Table';
-import TableBody from '../TableBody';
-import TableCell from '../TableCell';
-import TableHeader from '../TableHeader';
-import type { SortDirectionsType } from '../TableHeaderCell';
-import TableRow from '../TableRow';
+import { Filters } from './Filters';
+import { StackedCardsToTable } from './StackedCardsToTable';
+import { Table, type SortDirectionsType } from './Table';
+import { chromaticViewports } from '../../../src/util/viewports';
+
 import styles from './Table.stories.module.css';
 
 export default {
   title: 'Components/Table',
   component: Table,
-  subcomponents: { TableBody, TableCell, TableHeader, TableRow },
+  subcomponents: {
+    'Table.Body': Table.Body,
+    'Table.Caption': Table.Caption,
+    'Table.Cell': Table.Cell,
+    'Table.Header': Table.Header,
+    'Table.Row': Table.Row,
+  },
   parameters: {
     badges: ['1.1'],
   },
@@ -274,75 +279,102 @@ export const AlignTableCellContentRight: StoryObj<Args> = {
   },
 };
 
-const SortableExample = () => {
-  const values = [
-    { col1: 'Value 1', col2: 'Value A' },
-    { col1: 'Value 3', col2: 'Value B' },
-    { col1: 'Value 2', col2: 'Value C' },
-    { col1: 'Value 4', col2: 'Value D' },
-  ];
-  const [sortDirection, setSortDirection] =
-    useState<SortDirectionsType>('default');
-  const onSortClick = () => {
-    if (sortDirection === 'descending') {
-      setSortDirection('default');
-    }
-    if (sortDirection === 'default') {
-      setSortDirection('ascending');
-    }
-    if (sortDirection === 'ascending') {
-      setSortDirection('descending');
-    }
-  };
-  const sortedValues = values.slice().sort((a, b) => {
-    if (sortDirection === 'default') {
-      if (a.col2 < b.col2) {
-        return -1;
-      } else {
-        return 1;
+export const SortableInteractive: StoryObj<Args> = {
+  parameters: {
+    docs: {
+      source: {
+        type: 'dynamic',
+      },
+    },
+  },
+  render: () => {
+    const values = [
+      { col1: 'Value 1', col2: 'Value A' },
+      { col1: 'Value 3', col2: 'Value B' },
+      { col1: 'Value 2', col2: 'Value C' },
+      { col1: 'Value 4', col2: 'Value D' },
+    ];
+    const [sortDirection, setSortDirection] =
+      // eslint-disable-next-line react-hooks/rules-of-hooks
+      useState<SortDirectionsType>('default');
+    const onSortClick = () => {
+      if (sortDirection === 'descending') {
+        setSortDirection('default');
       }
-    }
-    if (sortDirection === 'ascending') {
-      if (a.col1 < b.col1) {
-        return -1;
-      } else {
-        return 1;
+      if (sortDirection === 'default') {
+        setSortDirection('ascending');
       }
-    }
-    if (sortDirection === 'descending') {
-      if (b.col1 < a.col1) {
-        return -1;
-      } else {
-        return 1;
+      if (sortDirection === 'ascending') {
+        setSortDirection('descending');
       }
-    }
-    return 0;
-  });
-  return (
-    <Table>
-      <Table.Header>
-        <Table.Row variant="header">
-          <Table.HeaderCell
-            onSortClick={onSortClick}
-            sortDirection={sortDirection}
-          >
-            Sortable
-          </Table.HeaderCell>
-          <Table.HeaderCell>Not sortable</Table.HeaderCell>
-        </Table.Row>
-      </Table.Header>
-      <TableBody>
-        {sortedValues.map(({ col1, col2 }) => (
-          <Table.Row key={`row-${col1}-${col2}`}>
-            <Table.Cell>{col1}</Table.Cell>
-            <Table.Cell>{col2}</Table.Cell>
+    };
+    const sortedValues = values.slice().sort((a, b) => {
+      if (sortDirection === 'default') {
+        if (a.col2 < b.col2) {
+          return -1;
+        } else {
+          return 1;
+        }
+      }
+      if (sortDirection === 'ascending') {
+        if (a.col1 < b.col1) {
+          return -1;
+        } else {
+          return 1;
+        }
+      }
+      if (sortDirection === 'descending') {
+        if (b.col1 < a.col1) {
+          return -1;
+        } else {
+          return 1;
+        }
+      }
+      return 0;
+    });
+    return (
+      <Table>
+        <Table.Header>
+          <Table.Row variant="header">
+            <Table.HeaderCell
+              onSortClick={onSortClick}
+              sortDirection={sortDirection}
+            >
+              Sortable
+            </Table.HeaderCell>
+            <Table.HeaderCell>Not sortable</Table.HeaderCell>
           </Table.Row>
-        ))}
-      </TableBody>
-    </Table>
-  );
+        </Table.Header>
+        <Table.Body>
+          {sortedValues.map(({ col1, col2 }) => (
+            <Table.Row key={`row-${col1}-${col2}`}>
+              <Table.Cell>{col1}</Table.Cell>
+              <Table.Cell>{col2}</Table.Cell>
+            </Table.Row>
+          ))}
+        </Table.Body>
+      </Table>
+    );
+  },
 };
 
-export const SortableInteractive: StoryObj<Args> = {
-  render: () => <SortableExample />,
+export const FiltersInteractive: StoryObj<Args> = {
+  parameters: {
+    badges: ['1.1', 'implementationExample'],
+  },
+  render: () => <Filters />,
+};
+
+export const StackedCardsExample: StoryObj<Args> = {
+  parameters: {
+    badges: ['1.1', 'implementationExample'],
+    chromatic: {
+      viewports: [
+        chromaticViewports.googlePixel2,
+        chromaticViewports.ipadMini,
+        chromaticViewports.chromebook,
+      ],
+    },
+  },
+  render: () => <StackedCardsToTable />,
 };